resolve discretização geográfica por bairros
